requirements.txt
#sudo apt install libmariadb3 libmariadb-dev
#sudo apt-get install ghostscript
Flask==2.2.3
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
bcrypt==4.0.1
python-dotenv==1.0.0
mariadb==1.1.6
PyJWT==2.6.0
flask-restx==1.1.0
Flask-Mail==0.9.1
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
boto3==1.26.129
openai-whisper==20230124
stable-ts==1.1.5
requests==2.30.0
PyMuPDF==1.22.3
Pillow==10.2.0
docx2txt==0.8
opencv-python==4.8.0.74
# Added - JLRC (03112023 -)
Werkzeug==2.3.3
#Beto
mlflow==2.9.2
fpdf==1.7.2
python-doctr==0.7.0
PyPDF2==3.0.1
torchvision==0.15.0
rapidfuzz==3.6.1
pdf2image==1.16.3
pyannote.audio==3.1.0
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pydub==0.25.1